Modernize JetCorrectionESChain
- Merge source and header, and move to plugins
- Use ESGetToken
- Use make_unique
- Remove unnecessary call to ChainedJetCorrector::clear()
