+++ JetMETCorrections/Modules/plugins/SealModule.cc
@@ -6,7 +6,6 @@
 #include "JetMETCorrections/Objects/interface/JetCorrector.h"
 #include "JetMETCorrections/Modules/interface/JetCorrectionESProducer.h"
 #include "JetMETCorrections/Modules/interface/JetCorrectionESSource.h"
-#include "JetMETCorrections/Modules/interface/JetCorrectionESChain.h"
 #include "JetMETCorrections/Modules/interface/JetCorrectionProducer.h"
 #include "JetMETCorrections/Modules/interface/JetResolutionESProducer.h"
 #include "JetMETCorrections/Algorithms/interface/LXXXCorrector.h"
@@ -42,8 +41,6 @@ DEFINE_FWK_MODULE(GenJetCorrectionProducer);
 
 typedef JetCorrectionProducer<BasicJet> BasicJetCorrectionProducer;
 DEFINE_FWK_MODULE(BasicJetCorrectionProducer);
-
-DEFINE_FWK_EVENTSETUP_MODULE(JetCorrectionESChain);
 
 DEFINE_FWK_EVENTSETUP_MODULE(JetResolutionESProducer);
 DEFINE_FWK_EVENTSETUP_MODULE(JetResolutionScaleFactorESProducer);
